% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova}
\alias{anova}
\title{function to run ANOVA test on proxzyme data}
\usage{
anova(df, stat.time = NULL, stat.dna = "ssDNA T")
}
\arguments{
\item{df}{a dataframe of data containing each condition minus background, avg, and sd
should be the same df used in plot_maker}

\item{stat.time}{anova test of data at that time point}

\item{stat.dna}{input for the dna condition of interest, default is "ssDNA T",
for another condition, name must be in "" in input.}
}
\description{
This function filters the processed data by time and dna condition,
creates a linear model of absorbance with respect to treatment, and then runs aov test.
}
