% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mg_and_edta_data.R
\docType{data}
\name{mg_and_edta_data}
\alias{mg_and_edta_data}
\title{Proxzyme Reaction with different concentrations of Mg with or wo EDTA}
\format{
a data frame containing 88 rows and 14 columns
\describe{
\item{time}{Time at which the absorbance was measured, in minutes}
\item{row}{Lettered rows on the 96 well plate, A-F}
\item{1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12}
{Numbered columns of the 96 well plate, 1-12}
}
}
\usage{
mg_and_edta_data
}
\description{
Absorbance reads from 1 to 32 minutes of Proxzyme reaxtion with TMB substrate.
There were four DNA conditions and eight buffer treatments/reaction conditions.
}
\keyword{datasets}
