% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mg_plate_key.R
\docType{data}
\name{mg_plate_key}
\alias{mg_plate_key}
\title{Plate Key for Mg and EDTA Experiment}
\format{
a data frame with 96 rows and 4 columns
\describe{
\item{row}{Lettered rows of the 96 well plate, A-H}
\item{col}{Numbered columns of the 96 well plate, 1-12}
\item{dna}{The DNA treatment for the corresponding well,
No DNA, contains no DNA, only buffer,
No Target, contains proxzyme DNA, but no target sequence,
P2 Only, contains only proxzyme 2, with ssDNA target
ssDNA T, contains both proxzymes, with ssDNA target}
\item{treatment}{the buffer treatment/reaction condition for the corresponding well,
water, contained no additional buffer components (TE buffer, *Salt, 5mg/50ml Hemin, DMSO),
IsoAMP, contains NEB Isothermal Amplification buffer II in addition to
standard buffer components,
0.1mM Mg, 0.2mM Mg, 0.5mM Mg, Mg concentrations without EDTA
0.1mM Mg+EDTA, 0.2mM Mg+EDTA, 0.5mM Mg+EDTA, Mg concentrations with equal mM EDTA}
}
}
\usage{
mg_plate_key
}
\description{
Key for plate layout for mg experiment, specifying the treatment/condition for each well
}
\keyword{datasets}
