% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minus_bg.R
\name{minus_bg}
\alias{minus_bg}
\title{function to subtract background based on whether there is a reference time}
\usage{
minus_bg(df, ref.time = NULL)
}
\arguments{
\item{df}{a dataframe of processed proxzyme data}

\item{ref.time}{time at which the rxn has progressed the most from the previous time point}
}
\description{
This function uses an if statement to choose which process to run based on
whether or not a reference time is assigned in the input. Either process will subtract the
No DNA condition from each treatment to remove background reaction and give absorbance due to
DNA.
}
\details{
The reference time must be determined beforehand by looking at the change in absorbance
between each timepoint. If there is no reference time, this function will subtract the average
of the No DNA condition for each treatment at each time from each replicate. When a reference
time is assigned, the function will subtract the average No DNA condition for each treatment at the reference time
from each replicate at every other time.
The output of that decision process is mutated to add the average and standard deviation
of each time, DNA, and treatment group to the data frame. This is done to prepare the data for
the next step in the analysis process--visualization.
}
