% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maker_new.R
\name{plot_maker}
\alias{plot_maker}
\title{function to plot proxzyme data}
\usage{
plot_maker(
  df,
  bar.chart = FALSE,
  plot.time = NULL,
  print.plot = TRUE,
  save.fn = NULL
)
}
\arguments{
\item{df}{a dataframe of data containing each condition minus background, avg, and sd}

\item{bar.chart}{if you want the bar chart, bar.chart = TRUE, leave blank for scatterplot}

\item{plot.time}{the time(s) that will be plotted by the function}

\item{save.fn}{the name of the plot image to be saved}
}
\description{
This function first filters a dataframe to remove all No DNA values, then
uses an if statement to let the user determine which time(s) will be plotted, then
uses another if statement to let the user determine if a bar chart or a scatterplot should
be made, and the last is statement give the user the option to assign a file name for the
plot to be saved.
}
\details{
bar.chart must be TRUE in order to generate a bar chart, default is FALSE and will
generate a scatterplot.
plot.time must be either a single numerical timepoint or in the form c(#, #, ...) for
multiple timepoints, the default is NULL and will generate a plot with all timepoints.
print.plot must be FALSE if the user does not wish to have the plot printed. The default
is TRUE.
save.fn mus be the name of the file, including the file type (.png, .jpg, etc.). This
should contain the path for the file if the user wants it to go to a different directory.
The default is NULL, so if no file name is given, the plot will not be saved.
}
