% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_proxzyme.R
\name{preprocess_proxzyme}
\alias{preprocess_proxzyme}
\title{function to read and pivot raw proxzyme absorbance data}
\usage{
preprocess_proxzyme(raw.fn, key.fn)
}
\arguments{
\item{raw.fn}{a csv file containing raw absorbance data from a 96 well plate}

\item{key.fn}{a csv file with the treatments in each well}
}
\description{
This function will read the raw data, convert it to long-form data,
read the plate key, and join the two in order to assign each absorbance value a DNA treatment
and a buffer treatment.
}
\details{
All excel formats must be in csv format and raw data file should have no more than
14 columns, with the first column being the time at which the reading was taken and the
second column being the lettered rows of the plate. The plate key must be made in Excel
in long form in order to have separate columns for DNA treatment and buffer/rxn condition.
}
