% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-tween_package_data.R
\docType{data}
\name{tween_package_data}
\alias{tween_package_data}
\title{Proxzyme Reaction with Various Tween Concentrations}
\format{
a data frame with 66 rows and 14 columns
\describe{
\item{time}{Time at which the absorbance was measured, in minutes}
\item{row}{Lettered rows on the 96 well plate, A-F}
\item{1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12}
{Numbered columns of the 96 well plate, 1-12}
}
}
\usage{
tween_package_data
}
\description{
Absorbance reads from 1 to 32 minutes of Proxzyme reaction with TMB substrate.
There were four DNA conditions and six buffer treatments/reaction conditions.
}
\keyword{datasets}
